feat(DelimiterValidationModel): 구분자에 따라 문자열 검증하는 함수 클래스 분리
